refactor(suite): move EverstakeModal to its own dir and update its content
