+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/EverstakeModal/EverstakeModal.tsx
