+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/UserContextModal.tsx
@@ -9,6 +9,8 @@ import { useDispatch } from 'src/hooks/suite';
 import type { AcquiredDevice } from 'src/types/suite';
 
 import { CardanoWithdrawModal } from '../CardanoWithdrawModal';
+import { ConfirmAddressModal } from '../ConfirmAddressModal';
+import { ConfirmXpubModal } from '../ConfirmXpubModal';
 import type { ReduxModalProps } from '../ReduxModal';
 import { AddAccountModal } from './AddAccountModal/AddAccountModal';
 import { AddTokenModal } from './AddTokenModal';
@@ -29,30 +31,28 @@ import { ConnectLoadingModal } from './ConnectLoadingModal';
 import { ConnectPermissionsModal } from './ConnectPermissionsModal';
 import { CopyAddressModal } from '../CopyAddressModal';
 import { CriticalCoinjoinPhaseModal } from './CriticalCoinjoinPhaseModal/CriticalCoinjoinPhaseModal';
+import { DeviceAuthenticityOptOutModal } from './DeviceAuthenticityOptOutModal';
 import { DisableTorModal } from './DisableTorModal';
 import { DisableTorStopCoinjoinModal } from './DisableTorStopCoinjoinModal';
+import { EverstakeModal } from './EverstakeModal/EverstakeModal';
 import { FirmwareRevisionOptOutModal } from './FirmwareRevisionOptOutModal';
 import { ImportTransactionModal } from './ImportTransactionModal/ImportTransactionModal';
 import { MetadataProviderModal } from './MetadataProviderModal/MetadataProviderModal';
 import { MoreRoundsNeededModal } from './MoreRoundsNeededModal';
 import { PinMismatchModal } from './PinMismatchModal';
+import { QrScannerModal } from './QrScannerModal';
 import { RequestEnableTorModal } from './RequestEnableTorModal';
 import { SafetyChecksModal } from './SafetyChecksModal';
 import { StakeInANutshellModal } from './StakeInANutshellModal/StakeInANutshellModal';
 import { StakeModal } from './StakeModal/StakeModal';
 import { TorLoadingModal } from './TorLoadingModal';
 import { TradingDCAModal } from './TradingDCAModal';
-import { TxSimulationModal } from './TxSimulationModal';
-import { EverstakeModal } from './UnstakeModal/EverstakeModal';
-import { WalletConnectProposalModal } from './WalletConnectProposalModal';
-import { WalletConnectSwitchAccountModal } from './WalletConnectSwitchAccountModal';
-import { ConfirmAddressModal } from '../ConfirmAddressModal';
-import { ConfirmXpubModal } from '../ConfirmXpubModal';
-import { DeviceAuthenticityOptOutModal } from './DeviceAuthenticityOptOutModal';
-import { QrScannerModal } from './QrScannerModal';
 import { TradingTermsModal } from './TradingTermsModal';
 import { TxDetailModal } from './TxDetailModal/TxDetailModal';
+import { TxSimulationModal } from './TxSimulationModal';
 import { UnecoCoinjoinModal } from './UnecoCoinjoinModal';
+import { WalletConnectProposalModal } from './WalletConnectProposalModal';
+import { WalletConnectSwitchAccountModal } from './WalletConnectSwitchAccountModal';
 import { PinInvalidModal } from '../DeviceContextModal/PinInvalidModal';
 import { TransactionReviewModal } from '../TransactionReviewModal/TransactionReviewModal';
 import { UnstakeModal } from './UnstakeModal/UnstakeModal';
